数据库里面的submission的得分是可以为小数的
排序好像正好是倒着的，已修正

+++ database/models/submission.js
@@ -13,7 +13,7 @@
         defaultValue: 'WT'
       },
       score: {
-        type: DataTypes.INTEGER
+        type: DataTypes.FLOAT
       },
       time_cost: {
         type: DataTypes.INTEGER
